in-game NPC-NPC communication prototype
Logic for NPC-NPC communication has been transferred to the environment of the simulation. Now, instead of pressing specific buttons to start a conversation, npcs react to physically  getting close to eachother. Since there is no real movement logic for the NPCs yet, the scene is scripted so one NPC will always approach another one, thus the conversation starts a few seconds after starting the program. Meanwhile, the user is still able to start a chat with another NPC of the scene, the two conversations won't interrupt eachother.