Interaction refactor and schedule base
Interaction class has been refactored to behave correctly in scaled environment. It also ensures now that NPCs already taking part in a conversation cannot start another one.

Objects are placed on the positions of the future buildings on the map with an identification script and a basic, currently manual daily schedule assignment had been given to the NPCs. The movement logic according to the schedule is not implemented yet.