Conversation finishing logic
Conversations are now finished automatically when two chatting npcs leave eachother.
Npc-user logic is also adjusted, so an NPC that is talking to the user will not abandon the conversation when the timer triggers its next place to be, instead it will stay still and only the player is able to stop the conversation.