Fallback text generator dummy
Backup implementation for cases when no LLM can be accessed. Instead of just printing out simple nonsense texts, the implementation contains a few whole sentences and a minimal logic to generate seemingly realistic conversations even without
 LLM, thus making it look better.
The main goal of the dummy is to make development of some features possible even without LLM and make the conversations a bit similar to the real LLM answers.
Although it is not the main purpose, but the logic might also come handy in cases when there is an accessible LLM but it sometimes has an outage because one-one message generated by this dummy may seem appropriate in the context (in contrast to just outputting numbers or simple words when the LLM call fails).